Fixed Cassandra Dynamic Query support with nameparams
